loadtest/requirements.txt
locust>=2.20
psycopg2-binary>=2.9
